+++ sereto/pdf.py
@@ -51,7 +51,7 @@ def render_report_pdf(settings: Settings, version: ReportVersion, recipe: str | 
     Returns:
         Path to the rendered PDF file.
     """
-    report_path = Report.get_path(dir_subtree=settings.reports_path)
+    report_path = Report.get_path_from_cwd(dir_subtree=settings.reports_path)
     report_tex_path = report_path / f"report{version.path_suffix}.tex"
 
     # Get the recipe to render the report
@@ -82,7 +82,7 @@ def render_sow_pdf(report: Report, settings: Settings, version: ReportVersion, r
         recipe: Name which will be used to pick a recipe from Render configuration. If none is provided, the first
             recipe (index 0) is used.
     """
-    report_path = Report.get_path(dir_subtree=settings.reports_path)
+    report_path = Report.get_path_from_cwd(dir_subtree=settings.reports_path)
     sow_tex_path = report_path / f"sow{version.path_suffix}.tex"
 
     if recipe is None:
@@ -111,7 +111,7 @@ def render_target_pdf(
         recipe: Name which will be used to pick a recipe from Render configuration. If none is provided, the first
             recipe (index 0) is used.
     """
-    report_path = Report.get_path(dir_subtree=settings.reports_path)
+    report_path = Report.get_path_from_cwd(dir_subtree=settings.reports_path)
     target_tex_path = report_path / f"{target.uname}.tex"
     replacements = {"%TARGET_DIR%": str(report_path / target.uname)}
 
@@ -147,7 +147,7 @@ def render_finding_group_pdf(
         recipe: Name which will be used to pick a recipe from Render configuration. If none is provided, the first
             recipe (index 0) is used.
     """
-    report_path = Report.get_path(dir_subtree=settings.reports_path)
+    report_path = Report.get_path_from_cwd(dir_subtree=settings.reports_path)
     finding_group_tex_path = report_path / f"{target.uname}_{finding_group.uname}.tex"
     replacements = {"%FINDINGS_DIR%": str(report_path / target.uname / "findings")}