+++ sereto/target.py
@@ -112,7 +112,7 @@ def render_target_j2(
     convert_recipe: str | None = None,
 ) -> None:
     cfg = report.config.at_version(version=version)
-    report_path = Report.get_path(dir_subtree=settings.reports_path)
+    report_path = Report.get_path_from_cwd(dir_subtree=settings.reports_path)
 
     render_target_findings_j2(target=target, settings=settings, version=version, convert_recipe=convert_recipe)